+++ examples/src/Fresh.hs
@@ -1,13 +1,14 @@
 {-# LANGUAGE NoImplicitPrelude #-}
 module Fresh (module Fresh) where
 
-import Data.Function (($), flip)
+import Data.Function (($), (.), flip)
 import Data.Monoid ((<>))
 import System.IO (IO)
 import Text.Show (show)
 
+import Control.Monad.Freer (runM)
 import Control.Monad.Freer.Fresh (evalFresh, fresh)
-import Control.Monad.Freer.Trace (runTrace, trace)
+import Control.Monad.Freer.Trace (runTraceIO, trace)
 
 
 -- | Generate two fresh values.
@@ -16,7 +17,7 @@ import Control.Monad.Freer.Trace (runTrace, trace)
 -- Fresh 0
 -- Fresh 1
 traceFresh :: IO ()
-traceFresh = runTrace $ flip evalFresh 0 $ do
+traceFresh = runM . runTraceIO $ flip evalFresh 0 $ do
     n <- fresh
     trace $ "Fresh " <> show n
     n' <- fresh